requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
tokenizers==0.20.3
datasets==3.2.0
scipy==1.14.1
google-cloud-storage==2.19.0
ipython==8.28.0
ipdb==0.13.13
torch>=2.1.0
torchdata==0.8.0
transformers>=4.36.0
wandb==0.19.1
pytest==8.3.5
einops>=0.7.0
numpy==1.26.4
typing-extensions>=4.5.0
typeguard==4.4.1
editdistance==0.8.1
accelerate>=0.27.0